Move classes to the kubeless.py file
